Add Post-Event Change-Check.
This patch takes care of important fixes for the placer duping blocks, the item replacer duping blocks, and the torch placer being able to place infinitely. It's not the source of the issue, but in any case for the future it will make sure if an exception has been thrown after a block is placed that it will still return true so that it can be removed from the player's / machine's inventory. 

Due to the fact that at least one of these was able to be reproduced in single player without sponge I would consider checking out what BlockState is being passed to the world. I think that the issue the exceptions are being thrown has to do with that, it is possible it could be missing some piece of information.