Fix a bug where stun and swimspeed potions were brewable from any potion
Because of how the Ingredient class works (it ignores NBT), any potion
(and not just the awkward potion) could be used to brew the potions of
this mod. This Commit fixes the bug by introducting a new class
ModBrewingRecipe which also checks the NBT of the input ItemStack (via
the PotionUtils.getPotionFromItem() method).

This commit also fixes an issue where the stun and swimspeed splash
potions were unbrewable because of a bug in the splashBrewing()
method in the PotionRegistry class.