Implement new Fireplace block. 2 behaviors with 2 variants: When it receives power, tries to a light a fire on the block in front of it if able to do so When it loses power, tries to extinguish fire in front of it
Variant behavior: If there is an infinitely-burnable block in front of it (such as netherrack), it will attempt to light the top of that block on fire or extinguish an existing fire on top of that block.

Also added infiniburn_overworld tag so the fireplace can light itself on fire when facing up, and stay lit.