Fix AutoCaveTorchItem to only place torches on valid surfaces
This commit changes the behavior of the AutoCaveTorchItem to be more
stringent regarding the placement of torches. This is done by using the
tryPlace() method of the BlockItem class, to make sure that the surface
is a valid placement for torches. This a fix (at least for one of the
mentioned problems) for issue #1801.